pyproject.toml
[project]
name = "guestpass-bot"
version = "0.1.0"
description = "Discord bot for automating guest parking registration with PPOA"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.4.0",
    "playwright>=1.48.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.3",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]

[tool.mypy]
files = ["src"]
mypy_path = ["src"]
explicit_package_bases = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Avoid specifying long messages outside the exception class
    "TRY003",
    # Consider moving statement to else block
    "TRY300",
]

[tool.ruff.format]
preview = true
